the end

+++ database/item.py
@@ -1,5 +1,4 @@
-import uuid
-from flask import request
+from flask_jwt_extended import jwt_required,get_jwt
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 
@@ -11,10 +10,10 @@ from sqlalchemy.exc import SQLAlchemyError,IntegrityError
 blp = Blueprint("items", __name__, description = "Operation on Items")
 
 
-
 @blp.route("/items")
 class ItemList(MethodView):
-
+  
+  @jwt_required(fresh=True)
   @blp.arguments(ItemSchema)
   @blp.response(201, ItemSchema)
   def post(self, request_data):
@@ -31,6 +30,8 @@ class ItemList(MethodView):
       
     return item, 201
   
+
+  
   @blp.response(200, ItemSchema(many=True))
   def get(self):
     item = ItemModel.query.all()
@@ -45,8 +46,11 @@ class Items(MethodView):
     item = ItemModel.query.get_or_404(item_id)
     return item
 
-
+  @jwt_required()
   def delete(self, item_id):
+    jwt = get_jwt()
+    if not jwt.get("is_admin"):
+      abort(401, message="Admin privilege required")
     item = ItemModel.query.get_or_404(item_id)
     db.session.delete(item)
     db.session.commit()
